Let boxes change color on mouse enter
Added an event listener to each box that looks for mouse enter. When the
mouse enters the box, it changes the background colour to black allowing
the user to draw on the grid.
Additionally removed margins and gaps from the grid so that the can be
whole making the game act more like "pen on paper" when the mouse moves
across the grid.