Add random grid cell color change on event
Modified the event listener for each cell in the grid to randomise the
RGB colour selection upon each mouse enter. This is to satisfy one of
the "extra credit" activities in the TOP project exercise.

Additionally added very basic styling to the button and header on the
webpage to make everything appear centered on the screen.